Committing one file at a time because I made too many changes without comitting. That means that any given version will likely not compile until I have every change committed.
Made SteamShortcuts enumerable for obvious reasons - it is a container for SteamShortcut.
Added ShortCutReplacePolicy for possible future changes where I allow the user to dictate what happens when there are duplicate shortcuts. This isn't fully implemented, and there's no way to ask the app to specify a policy yet.
Documented some functions.
Updated GenerateShortcut and GenerateAndAddShortcut to comply with constructor changes in SteamShortcut in the last commit.
Important - added Parsing of shortcuts.vdf into a SteamShortcuts collection with Parse, saving a collection to shortcuts.vdf with Save and printing the list to the console with Print. I might implement ToString and have print just print the string to the console in a future commit.