Committing one file at a time because I made too many changes without comitting. That means that any given version will likely not compile until I have every change committed. This is the sixth commit.
Tiny change to spacing on Emulator.Expand. Barely worth mentioning.

Added TitlePattern to the EmulatorConfiguration (and thus support for that in the emulators.ini file).

Added a butt-ton of documentation inside of the emulators.ini file, giving usage examples and explanations of every option.

Fixed retroarch parameters to not use --menu when launching a rom.

Added CLI using the CommandLineParser library.

Added CommandLineParser and  Crc32C.NET packages.