Added %d substitution variable. %d does what %D did before. Changed %D to get the parent directory's absolute path. Added method to calculate that the %D path.
Added category attributes for Icons, which are separate from attributes for images. Icons will be actual ico files, or exes with icons in them if you use the comma + index at the end. Images will be used for full images like jpegs and pngs. They'll be used as the steam grid image, while the icon will be the steam icon, duh. Icon will also be used for windows shortcuts.