Added commented out versions of parameters to use snes9x.
Removed sample code for using my script to call cemu. Too specific to me to include in a sample config. I should keep the sample more generic.

Samples for NDS with either retroarch or desmume. Desmume sucks for using in fullscreen with a controller, but is a better emulator otherwise, while retroarch is perfect for using with steam or other launchers.

Updated help text referring to the %p substitution, since it was changed to be relative to startin.